Clean up callback handling (#31)
* mid-foolery

* Cleaned up password handling and split archive delegates.  Also cleaned up Aes code.
