+++ src/EggDotNet.Tests/UnitTest1.cs
@@ -119,10 +119,27 @@ namespace EggDotNet.Tests
 		}
 
 		[Fact]
+		public void Test_ZipEnc()
+		{
+			using var fs = new FileStream(GetTestPath("lorem_long_zipEnc.egg"), FileMode.Open, FileAccess.Read);
+			using var archive = new EggArchive(fs, false, null, (filename, options) => { options.Password = "password12345!"; options.Retry = false; });
+			Assert.Equal("Lorem long text encrypted with ZIP", archive.Comment);
+			ValidateAllEggEntries(archive);
+			var aes256Entry = archive.GetEntry("lorem_ipsum_long.txt");
+			Assert.Equal("This file is encrypted using ZIP.", aes256Entry!.Comment);
+			using var entryStream = aes256Entry.Open();
+			using var sr = new StreamReader(entryStream);
+			var loremLongText = sr.ReadToEnd();
+			Assert.Equal(15_238, loremLongText.Length);
+			Assert.StartsWith("Lorem ipsum dolor sit amet", loremLongText);
+			Assert.EndsWith("sed faucibus orci ligula eu nisi.", loremLongText);
+		}
+
+		[Fact]
 		public void Test_Aes128()
 		{
 			using var fs = new FileStream(GetTestPath("lorem_long_aes128.egg"), FileMode.Open, FileAccess.Read);
-			using var archive = new EggArchive(fs, false, null, () => "password12345!");
+			using var archive = new EggArchive(fs, false, null, (filename, options) => { options.Password = "password12345!"; options.Retry = false; });
 			Assert.Equal("Lorem long text encrypted with AES128", archive.Comment);
 			ValidateAllEggEntries(archive);
 			var aes256Entry = archive.GetEntry("lorem_ipsum_long.txt");
@@ -139,7 +156,7 @@ namespace EggDotNet.Tests
 		public void Test_Aes256()
 		{
 			using var fs = new FileStream(GetTestPath("lorem_long_aes256.egg"), FileMode.Open, FileAccess.Read);
-			using var archive = new EggArchive(fs, false, null, () => "password12345!");
+			using var archive = new EggArchive(fs, false, null, (filename, options) => { options.Password = "password12345!"; options.Retry = false; });
 			Assert.Equal("Lorem long text encrypted with AES256", archive.Comment);
 			ValidateAllEggEntries(archive);
 			var aes256Entry = archive.GetEntry("lorem_ipsum_long.txt");
