+++ src/EggDotNet/Encryption/Aes/CryptoMode.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace EggDotNet.Encryption.Aes
+{
+	internal enum CryptoMode
+	{
+		Encrypt,
+		Decrypt
+	}
+}
