+++ src/EggDotNet/Encryption/IStreamDecryptionProvider.cs
@@ -4,7 +4,7 @@ namespace EggDotNet.Encryption
 {
 	internal interface IStreamDecryptionProvider
 	{
-		bool PasswordValid { get; }
+		bool AttachAndValidatePassword(string password);
 
 		Stream GetDecryptionStream(Stream stream);
 	}
