+++ src/EggDotNet/Encryption/Zip/ZipDecryptStream.cs
@@ -132,6 +132,4 @@ namespace EggDotNet.Encryption.Zip
 			throw new NotImplementedException();
 		}
 	}
-
-
 }
