+++ src/EggDotNet/Exceptions/DecryptFailedException.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace EggDotNet.Exceptions
+{
+	/// <summary>
+	/// Exception indicating that decryption has failed.
+	/// </summary>
+	public sealed class DecryptFailedException : Exception
+	{
+		internal DecryptFailedException()
+			: base("Decryption of entry failed")
+		{ }
+	}
+}
