+++ src/EggDotNet/Format/EggFileFormatFactory.cs
@@ -4,12 +4,13 @@ using EggDotNet.Format;
 using System.Collections.Generic;
 using System;
 using System.IO;
+using static EggDotNet.Callbacks;
 
 namespace EggDotNet.Format
 {
 	internal static class EggFileFormatFactory
 	{
-		public static EggFileFormatBase Create(Stream stream, Func<Stream, IEnumerable<Stream>> streamCallback, Func<string> pwCallback)
+		public static EggFileFormatBase Create(Stream stream, SplitFileReceiverCallback streamCallback, FileDecryptPasswordCallback pwCallback)
 		{
 #if NETSTANDARD2_1_OR_GREATER
 			Span<byte> headerBuffer = stackalloc byte[Global.HEADER_SIZE];
